feat: introduce `Script` value object for .json file scripts section

+++ tests/Unit/Services/FileEditors/JsonFileEditorTest.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 use Illuminate\Filesystem\Filesystem;
 use Techieni3\StacktifyCli\Services\FileEditors\JsonFileEditor;
+use Techieni3\StacktifyCli\ValueObjects\Script;
 
 $destinationDirectory = dirname(__DIR__).'/../../Workspace';
 
@@ -15,7 +16,9 @@ beforeEach(function () use ($destinationDirectory): void {
 });
 
 afterEach(function () use ($destinationDirectory): void {
-    unlink($destinationDirectory.'/composer.json');
+    if (file_exists($destinationDirectory.'/composer.json')) {
+        unlink($destinationDirectory.'/composer.json');
+    }
 });
 
 it('returns false when saving without changes', function () use ($destinationDirectory): void {
@@ -27,9 +30,26 @@ it('returns false when saving without changes', function () use ($destinationDir
 it('adds a new script', function () use ($destinationDirectory): void {
     $composerJson = new JsonFileEditor($destinationDirectory.'/composer.json');
 
-    $composerJson->addScript('test', 'php artisan test');
+    $composerJson->addScript(new Script('test', 'php artisan test'));
 
     expect($composerJson->save())->toBeTrue()
         ->and(file_get_contents($destinationDirectory.'/composer.json'))->toContain('"test": "php artisan test"');
 
+});
+
+it('adds a script with array command', function () use ($destinationDirectory): void {
+    $composerJson = new JsonFileEditor($destinationDirectory.'/composer.json');
+
+    $command = [
+        'npm run dev',
+        'npm run build',
+    ];
+    $composerJson->addScript(new Script('scripts', $command));
+
+    expect($composerJson->save())->toBeTrue();
+
+    $content = file_get_contents($destinationDirectory.'/composer.json');
+
+    expect($content)->toContain('"npm run dev"')
+        ->and($content)->toContain('"npm run build"');
 });
